requirements.txt
streamlit>=1.28.0
sentence-transformers>=2.2.2
chromadb>=0.4.15
newspaper3k>=0.2.8
requests>=2.31.0
pandas>=2.0.0
scikit-learn>=1.3.0
beautifulsoup4>=4.12.0
feedparser>=6.0.10
lxml[html_clean]>=4.9.0
transformers>=4.30.0
torch>=2.0.0
openai>=1.0.0
python-dotenv>=1.0.0
google-generativeai>=0.3.0